fix: correct mapping of ags/rs
Previously, all ags/rs codes were mapped to 'NAP_MajorRoad.allInLAUCode'.
This is changes here so that depending if the source value is an 'ags' or 'rs'
from a 'Gemeinde', 'Verbandsgemeinde'/'Amt' or 'Regierungsbezirk' the
attribute 'NAP_MajorRoad.allInLAUCode' or 'NAP_MajorRoad.allInNUTSCode' on target
side is filled correctly.

SVC-1696